# perf(python): optimize `_unpack_schema()`