Provides significant performance improvements to `pl.utils._construction._unpack_schema()`:

1. `python -m timeit -s "import polars as pl" "pl.utils._construction._unpack_schema(None)"`
487 nsec vs. 116 nsec per loop -> _4.20x_

1. `python -m timeit -s "import polars as pl; schema = tuple(f'col_{i}' for i in range(10))" "pl.utils._construction._unpack_schema(schema)"`
1.04 usec vs. 838 nsec per loop -> _1.24x_

1. `python -m timeit -s "import polars as pl; schema = {f'col_{i}': pl.Float64 for i in range(10)}" "pl.utils._construction._unpack_schema(schema)"`
2.82 usec vs. 1.85 usec per loop -> _1.52x_

1. `python -m timeit -s "import polars as pl; n_expected = 10; schema_overrides = {f'col_{i}': pl.Float64 for i in range(n_expected)}" "pl.utils._construction._unpack_schema(None, schema_overrides=schema_overrides, n_expected=n_expected)"`
3.44 usec vs. 2.13 usec per loop -> _1.15x_

1. ...